Fix NPM package smoke testing
If you work from the repository root, the latest version of fauxauth is
already installed - you test the local code, which means rebuilding the
lib/ outputs, not the deployed package.

Instead, this runs directly in packages/e2e/, which means using semver
in package.json not exact versions in package-lock.json.